Move the soft validation logic into a separate class
KAFKA-277
